COC-83-refactor(#79): api 단일 책임으로 만들고 핸들러에 로직 이동
