+++ src/api/domain/study.ts
@@ -1,7 +1,6 @@
 import { axiosInstance } from '@api/axiosInstance';
 import { END_POINTS_V1 } from '@constants/api';
 import { CreateStudyData, EditStudyData, GetListParams } from '@customTypes/study';
-import { SpaceData, SpaceListParams } from '@customTypes/space';
 
 const studyApi = {
   getInfo: async (studyId: string) => {
@@ -36,26 +35,6 @@ const studyApi = {
   edit: async (studyId: string, editStudyData: EditStudyData) => {
     const { data } = await axiosInstance.post(END_POINTS_V1.STUDY.EDIT(studyId), editStudyData);
     return data.result;
-  },
-
-  getSpaceList: async (studyId: string, params?: SpaceListParams): Promise<SpaceData[]> => {
-    const queryParams = params ? { ...params } : {};
-
-    Object.keys(queryParams).forEach((key) => {
-      if (queryParams[key] === null || queryParams[key] === undefined) delete queryParams[key];
-    });
-
-    const { data } = await axiosInstance.get(END_POINTS_V1.STUDY.SPACE_LIST(studyId), {
-      params: Object.keys(queryParams).length > 0 ? queryParams : undefined,
-      /* headers: {
-        Authorization: `Bearer ${token}`,
-      }, */
-    });
-
-    const startIndex = params?.lastIndex ?? 0;
-    const limit = 20;
-
-    return data.result.slice(startIndex, startIndex + limit);
   },
 };
