+++ src/mocks/handlers/study.ts
@@ -7,8 +7,6 @@ import {
   createStudyErrorResponse,
   editStudyErrorResponse,
   editStudyResponse,
-  getSpaceListErrorResponse,
-  getSpaceListResponse,
   getStudyInfoErrorResponse,
   getStudyInfoResponse,
   getStudyListResponse,
@@ -106,20 +104,6 @@ export const studyHandlers = [
       });
     }
     return new HttpResponse(JSON.stringify(editStudyResponse), {
-      status: HTTP_STATUS_CODE.SUCCESS,
-    });
-  }),
-
-  http.get(`${BASE_URL}${END_POINTS_V1.STUDY.SPACE_LIST(':studyId')}`, ({ params }) => {
-    const { studyId } = params;
-
-    if (!studyId || Number.isNaN(Number(studyId))) {
-      return new HttpResponse(JSON.stringify(getSpaceListErrorResponse), {
-        status: HTTP_STATUS_CODE.BAD_REQUEST,
-      });
-    }
-
-    return new HttpResponse(JSON.stringify(getSpaceListResponse), {
       status: HTTP_STATUS_CODE.SUCCESS,
     });
   }),
